+++ rabbit_monitor.py
@@ -29,6 +29,8 @@ Settings = namedtuple('Settings',
                           'rabbit_default_pass',
                           'rabbit_default_vhost',
                           'wait_time',
+                          'stats_window',
+                          'stats_incr',
                       ])
 
 settings = Settings(port=os.getenv("PORT", 5000),
@@ -36,7 +38,9 @@ settings = Settings(port=os.getenv("PORT", 5000),
                     rabbit_url=settings.RABBIT_URL,
                     rabbit_default_user=settings.RABBITMQ_DEFAULT_USER,
                     rabbit_default_pass=settings.RABBITMQ_DEFAULT_PASS,
-                    rabbit_default_vhost=settings.RABBITMQ_DEFAULT_VHOST)
+                    rabbit_default_vhost=settings.RABBITMQ_DEFAULT_VHOST,
+                    stats_window=settings.RABBIT_MONITOR_STATS_WINDOW,
+                    stats_incr=settings.RABBIT_MONITOR_STATS_INCREMENT,)
 
 healthcheck_url = settings.rabbit_url + 'healthchecks/node'
 aliveness_url = (settings.rabbit_url +
@@ -53,13 +57,18 @@ urls = {'healthcheck': healthcheck_url,
         'nodes': nodes_url,
         }
 
+url_parameters = {'lengths_age': settings.stats_window,
+                  'lengths_incr': settings.stats_incr,
+                  }
+
 
 @asyncio.coroutine
 def fetch(session, url):
     with aiohttp.Timeout(5):
         resp = None
         try:
-            return (yield from session.get(url))
+            return (yield from session.get(url,
+                                           params=url_parameters))
         except Exception as e:
             logger.error(e, status='bad')
             if resp is not None: