+++ settings.py
@@ -14,3 +14,12 @@ RABBIT_URL = 'http://{hostname}:{port}/api/'.format(
 )
 
 WAIT_TIME = os.getenv('RABBIT_MONITOR_WAIT_TIME', 300)
+
+# Number of seconds to look back and gather stats from
+RABBIT_MONITOR_STATS_WINDOW = os.getenv('RABBIT_MONITOR_STATS_WINDOW',
+                                        WAIT_TIME)
+
+# Sample frequency in stats window
+RABBIT_MONITOR_STATS_INCREMENT = os.getenv('RABBIT_MONITOR_STATS_INCREMENT',
+                                           str(RABBIT_MONITOR_STATS_WINDOW / 10),
+                                           )