+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add healthcheck endpoint for service
 - Add asynchronous behaviour for client/server actions
 - Add logging of message stats
+- Add stats time window to polling
 
 ### 0.1.0 2017-03-29
 - Poll url healthcheck and aliveness endpoints